Add check for number of replication
These changes provide a check number of replications. It groups the
units into an application and obtains the minimum number of
replications from the relation of the ceph-osd and ceph-mon
applications. The idea is to find the minimum number from the
replication list calculated from the pools "size" - "min_size".

If no pools are available, the check returns the Result(True).

Output example: [1]

---
[1]: https://pastebin.ubuntu.com/p/8W4YyBVj6d/