removed `BaseVerifier.run_action_on_units`
The function `BaseVerifier.run_action_on_units` was used with restriction
that prevents actions being run on units other than those that are part
of the self.units.

Because this restriction was no longer useful, it has been removed, and
without it, there is no reason for the `run_action_on_units` function
to be part of the `BaseVerifier`. At the same time, the
`run_action_on_unit` function has been moved from `BaseVerifier` to
`utils.unit`.